fix: make i2c device access reliable
Problem: On first request within a container i2c devices (/dev/i2c-1) are not visible since their creation event is not própagated into the container even if the container is run with `--privileged`.
Solution: mount the entire /dev dir into the container (would be a problem only if we'd run the container it -it).